[deleted messages] Refactor the constant to be able to give a reason
