+++ tests/message/test_no_removed_msgid_or_symbol_used.py
@@ -2,8 +2,8 @@
 # For details: https://github.com/PyCQA/pylint/blob/main/LICENSE
 # Copyright (c) https://github.com/PyCQA/pylint/blob/main/CONTRIBUTORS.txt
 
-from pylint.constants import DELETED_MESSAGES
 from pylint.lint import PyLinter
+from pylint.message.deleted_message_ids import _DELETED_MESSAGES_IDS
 
 
 def test_no_removed_msgid_or_symbol_used(linter: PyLinter) -> None:
@@ -12,7 +12,8 @@ def test_no_removed_msgid_or_symbol_used(linter: PyLinter) -> None:
     This could cause occasional bugs, but more importantly confusion and inconsistencies
     when searching for old msgids online. See https://github.com/PyCQA/pylint/issues/5729
     """
-    for msgid, symbol, old_names in DELETED_MESSAGES:
-        linter.msgs_store.message_id_store.register_message_definition(
-            msgid, symbol, old_names
-        )
+    for deleted_messages in _DELETED_MESSAGES_IDS.values():
+        for msgid, symbol, old_names in deleted_messages:
+            linter.msgs_store.message_id_store.register_message_definition(
+                msgid, symbol, old_names
+            )
